refactor: supabase리팩토링에 따라 영향을 받는 파일들 수정
- import 경로 변경
- db의 타입과 일치하지 않았던 타입들 수정
