+++ src/hooks/api/supabase/content-approvals/useGetApproveUser.ts
@@ -1,17 +1,12 @@
 import { getApproveUser } from '@/api/stories/api';
 import { QUERY_KEY } from '@/constants/queryKey';
-import { DBContentApprovalResponse } from '@/types/dbStory';
-import { useQuery, UseQueryResult } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 
 interface UseGetApproveUserParams {
   contentId?: string;
 }
 
-const useGetApproveUser = ({
-  contentId,
-}: UseGetApproveUserParams): UseQueryResult<
-  DBContentApprovalResponse[] | null
-> => {
+const useGetApproveUser = ({ contentId }: UseGetApproveUserParams) => {
   return useQuery({
     queryKey: [QUERY_KEY.GET_APPROVE_USER, contentId],
     queryFn: () => getApproveUser({ contentId: contentId! }),
