+++ src/hooks/api/supabase/contents/useGetLastContent.ts
@@ -1,7 +1,6 @@
 import { getLastContent } from '@/api/stories/api';
 import { QUERY_KEY } from '@/constants/queryKey';
-import { DBContentResponse } from '@/types/dbStory';
-import { useQuery, UseQueryResult } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 
 interface UseGetLastContentParams {
   storyId: string;
@@ -9,7 +8,7 @@ interface UseGetLastContentParams {
 
 const useGetLastContent = ({
   storyId,
-}: UseGetLastContentParams): UseQueryResult<DBContentResponse | null> => {
+}: UseGetLastContentParams) => {
   return useQuery({
     queryKey: [QUERY_KEY.GET_LAST_CONTENT, storyId],
     queryFn: () => getLastContent(storyId),
