+++ src/hooks/api/supabase/story-collaborators/useGetSummary.ts
@@ -1,18 +1,12 @@
 import { getSummary } from '@/api/social-detail/api';
 import { QUERY_KEY } from '@/constants/queryKey';
-import { useQuery, UseQueryResult } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 
 interface UseGetSummaryParams {
   socialId: number;
 }
 
-interface GetSummaryResponse {
-  summary: string;
-}
-
-const useGetSummary = ({
-  socialId,
-}: UseGetSummaryParams): UseQueryResult<GetSummaryResponse | null> => {
+const useGetSummary = ({ socialId }: UseGetSummaryParams) => {
   return useQuery({
     queryKey: [QUERY_KEY.GET_SUMMARY, socialId],
     queryFn: () => getSummary({ socialId: socialId }),
